Move testing utility methods up to mocks mod
